Extract ledger recovery and rollout helpers
