+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to Parsek are documented here.
 
 ### Internals
 
+- Continued refactor-4 with a behavior-neutral `LedgerOrchestrator` extraction for vessel recovery-funds pairing and rollout cost/adoption helpers. Existing orchestrator wrappers, constants, log tags/text, dedup formats, KSC sequence allocation, and build-cost residual emission stay in place while the moved logic lives in focused helper classes.
 - Continued refactor-4 (Pass 3) with a behavior-neutral `PostWalkActionReconciler` extraction: post-walk reconciliation classification, live-coverage filtering, event matching, coalesced aggregation, comparison, and dump diagnostics now live behind unchanged `LedgerOrchestrator` wrappers while shared commit-window helpers and de-duplication state stay in `LedgerOrchestrator`.
 - Continued refactor-4 (Pass 2) with a behavior-neutral `RecordingSidecarStore` save-path extraction: save-side path resolution, sidecar epoch bump/rollback, staged authoritative sidecar writes, readable mirror reconciliation, and `FilesDirty` clearing now live behind `RecordingStore` wrappers while load-path hydration stays in `RecordingStore`.
 - Continued refactor-4 (Pass 2) with a behavior-neutral `RecordingSidecarStore` load-path extraction: trajectory probe/id/epoch validation, sidecar load-failure marking, post-hydration loop and endpoint repairs, terminal-orbit backfill, and snapshot fallback policy now live behind `RecordingStore` wrappers while codecs stay in their existing owners.
